GappedReads-class {GenomicAlignments}	R Documentation

(Legacy) GappedReads objects

Description

The GappedReads class extends the GAlignments class.

A GappedReads object contains all the information contained in a GAlignments object plus the sequences of the queries. Those sequences can be accessed via the qseq accessor.

Constructor

GappedReads objects are typically created when reading a file containing aligned reads with the readGappedReads function.

Accessors

In the code snippets below, x is a GappedReads object.

qseq(x): Extracts the sequences of the queries as a DNAStringSet object.

Author(s)

Hervé Pagès

References

http://samtools.sourceforge.net/

See Also

GAlignments objects.

readGappedReads.

Examples

greads_file <- system.file("extdata", "ex1.bam", package="Rsamtools")
greads <- readGappedReads(greads_file)
greads
qseq(greads)

[Package GenomicAlignments version 1.26.0 Index]