setops-methods {GenomicAlignments}	R Documentation

Set operations on GAlignments objects

Description

Performs set operations on GAlignments objects.

NOTE: The pintersect generic function and method for IntegerRanges objects is defined and documented in the IRanges package. Methods for GRanges and GRangesList objects are defined and documented in the GenomicRanges package.

Usage

## S4 method for signature 'GAlignments,GRanges'
pintersect(x, y, ...)
## S4 method for signature 'GRanges,GAlignments'
pintersect(x, y, ...)

Arguments

x, y
A GAlignments object and a GRanges object. They must have the same length.

...
Further arguments to be passed to or from other methods.

Value

A GAlignments object parallel to (i.e. same length as) x and y.

See Also

The GAlignments class.

The setops-methods man page in the GenomicRanges package.

Examples

## Parallel intersection of a GAlignments and a GRanges object:
bamfile <- system.file("extdata", "ex1.bam", package="Rsamtools")
gal <- readGAlignments(bamfile)
pintersect(gal, shift(as(gal, "GRanges"), 6L))

[Package GenomicAlignments version 1.26.0 Index]